+++ src/main/java/fr/maxlego08/essentials/buttons/vault/admin/ButtonVaultOpenAdmin.java
@@ -23,7 +23,7 @@ public class ButtonVaultOpenAdmin extends Button {
     }
 
     @Override
-    public ItemStack getCustomItemStack(Player player) {
+    public ItemStack getCustomItemStack(Player player, Placeholders placeholders) {
         PlayerVaults viewerVaults = this.plugin.getVaultManager().getPlayerVaults(player);
         PlayerVaults targetVaults = viewerVaults.getTargetPlayerVaults();
         if (targetVaults == null) return this.getItemStack().build(player, false, new Placeholders());
@@ -32,7 +32,6 @@ public class ButtonVaultOpenAdmin extends Button {
         Vault currentVault = viewerVaults.getTargetVault();
 
         var itemstack = this.getItemStack();
-        Placeholders placeholders = new Placeholders();
         var vaultManager = plugin.getVaultManager();
 
         var vaultItemStack = targetVault.getIconItemStack();